post-review fixes
