+++ apps/projects/app/utils/work-ratings.js
@@ -1,0 +1,10 @@
+const workRatings = [
+    'Select a Rating',
+    '1 - Unusable',
+    '2 - Needs Rework',
+    '3 - Acceptable',
+    '4 - Exceeds Expectations',
+    '5 - Excellent',
+  ]
+
+export default workRatings
